Add GetStates method to IRecordsService interface
A new `using` directive was added to import the
`GymAssistant_API.Req_Res.Response.Progress` namespace in the
`IRecordsService.cs` file. Additionally, a new asynchronous method
`GetStates` was introduced in the `IRecordsService` interface. This
method returns a `Task<Result<StatesRes>>` and accepts a `userId`
(string) and an optional `CancellationToken` parameter.